fixes a number of issues with the Qt viewer:
* when Shift-RMB dragging, no rectangle would be overpainted in the viewer. Shift-RMB dragging draws a frame that will be zoomed, so this is really important...
* Shift select was not implemented, which allows for multiple selection

These changes allow for overpainting of the OCC OpenGL viewer. This paves the way for really cool widget to be drawn on top of the OCC OpenGL viewer. OCC has a API for painting a layer on top of the viewer, but its horrible and does not allow for widgets / interactivity.

An example is contributed, "examples/core_visualization_overpaint_viewer.py"

This example is WIP -- when rotating the view, then there's an issue syncing the buffers.
It looks like overpainting is processed later than the rotation of the viewport

Close to getting there though...